+++ lib/td_bg/search/search.ex
@@ -11,14 +11,17 @@ defmodule TdBg.Search do
   @index "concepts"
 
   def search(query) do
-    response = Elasticsearch.post(Cluster, "/#{@index}/_search", query)
+    response =
+      Elasticsearch.post(Cluster, "/#{@index}/_search", query,
+        params: %{"track_total_hits" => "true"}
+      )
 
     case response do
       {:ok, %{"aggregations" => aggregations, "hits" => %{"hits" => results, "total" => total}}} ->
-        %{results: results, total: total, aggregations: aggregations}
+        %{results: results, total: get_total(total), aggregations: aggregations}
 
       {:ok, %{"hits" => %{"hits" => results, "total" => total}}} ->
-        %{results: results, total: total, aggregations: %{}}
+        %{results: results, total: get_total(total), aggregations: %{}}
 
       {:error, %Elasticsearch.Exception{message: message} = error} ->
         Logger.warn("Error response from Elasticsearch: #{message}")
@@ -79,4 +82,7 @@ defmodule TdBg.Search do
   defp get_domain(""), do: nil
   defp get_domain(id) when is_integer(id) or is_binary(id), do: TaxonomyCache.get_domain(id)
   defp get_domain(_), do: nil
+
+  defp get_total(value) when is_integer(value), do: value
+  defp get_total(%{"relation" => "eq", "value" => value}) when is_integer(value), do: value
 end